requirements.txt
langchain==0.0.246
python-dotenv==1.0.0
gradio==3.27.0
gradio_client==0.1.3
transformers==4.28.0
tiktoken==0.3.3
openai==0.27.4
pypdf==3.13.0
chromadb==0.4.3
torch==2.0.0
lark==1.1.7